rfb: Set proper type for RFB_ENCODING_* to eliminate implicit sign conversion
src/rfb.c:269:8: error: implicit conversion changes signedness: 'int' to 'unsigned int' [-Werror,-Wsign-conversion]
                case RFB_ENCODING_RESIZE:
                     ^~~~~~~~~~~~~~~~~~~
src/rfb.c:135:30: note: expanded from macro 'RFB_ENCODING_RESIZE'
                                        ^~~~
1 error generated.

+++ src/rfb.c
@@ -130,9 +130,9 @@ struct rfb_pixfmt_msg {
 	struct rfb_pixfmt	pixfmt;
 };
 
-#define	RFB_ENCODING_RAW		0
-#define	RFB_ENCODING_ZLIB		6
-#define	RFB_ENCODING_RESIZE		-223
+#define	RFB_ENCODING_RAW		((uint32_t)0)
+#define	RFB_ENCODING_ZLIB		((uint32_t)6)
+#define	RFB_ENCODING_RESIZE		((uint32_t)-223)
 
 #define RFB_MAX_WIDTH			2000
 #define RFB_MAX_HEIGHT			1200
